Move filter sampling to a separate file
